image problems solved,error msgs added,cnfrm pswd vldn,checking ctgry

+++ routes/category-router.js
@@ -1,7 +1,9 @@
 const express = require('express');
 const categoryRouter = express.Router();
 const Category = require('../models/categoryModel');
+const Product = require('../models/productModel');
 
+const fs = require('fs');
 const auth = require('../config/auth');
 const { check, validationResult } = require('express-validator');
 
@@ -22,12 +24,17 @@ const upload = multer({storage : storage})
 
 
 categoryRouter.get('/',auth.isAdmin, (req,res)=>{
+    let count;
+    Category.count((err,c)=>{
+        count = c;
+    })
     Category.find( (err,categories)=>{
         if (err) return console.log(err);
-        // const message = req.flash('message')
         const admin = req.session.admin;
+        const success = req.flash('success');
+        const error = req.flash('error');
 
-        res.render('admin/category',{categories : categories,admin});
+        res.render('admin/category',{categories : categories,count:count,admin,success:success,error});
 
     });
 
@@ -37,22 +44,32 @@ categoryRouter.get('/add-category',auth.isAdmin,(req,res)=>{
 
     const admin = req.session.admin;
     const title = ""
-    res.render('admin/add-category',{admin,title});
+    const error = req.flash('error');
+
+    res.render('admin/add-category',{admin,title,error});
 });
 
 categoryRouter.post('/add-category',upload.single('image'), function (req, res) {
  
     let title = req.body.title.replace(/\s+/g, '-').toUpperCase();
     let slug = title.replace(/\s+/g, '-').toLowerCase();
-    let image = req.file.filename;
+    let image = typeof req.file !== "undefined" ? req.file.filename : "";
 
      
-        Category.findOne({slug}, function (err, category) {
+        Category.findOne({slug:slug}, function (err, category) {
+            if (err){
+                console.log('error in cat find');
 
+                return console.log(err);
+            }
             if (category) {
-                // req.flash('danger', 'Category title exists, choose another.');
                 console.log("cat exists");
-                res.redirect('/admin/category/add-category');
+                fs.unlink('public/images/category-img/'+ image ,(err)=>{
+                    if(err) console.log(err);
+                    console.log('old img deleted');
+                });
+                req.flash('error', 'Category title exists, choose another.');
+                return res.redirect('/admin/category/add-category');
             } else {
                 let category = new Category({
                     title: title,
@@ -83,7 +100,7 @@ categoryRouter.post('/add-category',upload.single('image'), function (req, res) 
                 });
             }
         
-         req.flash('success', 'Category added!');
+         req.flash('success', 'Category added successfully!');
                     res.redirect('/admin/category');
     
 
@@ -94,8 +111,9 @@ categoryRouter.get('/edit-category/:id',auth.isAdmin,(req,res)=>{
     Category.findById(req.params.id, (err,category)=>{
         if(err) return console.log(err);
         const admin = req.session.admin;
+        const error = req.flash('error');
 
-        res.render('admin/edit-category',{admin,
+        res.render('admin/edit-category',{admin,error:error,
             title: category.title,
             image : category.image,
             id : category._id
@@ -103,18 +121,41 @@ categoryRouter.get('/edit-category/:id',auth.isAdmin,(req,res)=>{
     })
 })
 
-categoryRouter.post('/edit-category/:id',(req,res)=>{
+categoryRouter.post('/edit-category/:id',upload.single('image'),(req,res)=>{
     const title = req.body.title;
-    const image = req.file.filename;
+    let image = typeof req.file !== "undefined" ? req.file.filename : "";
     const id = req.params.id;
+    const pimage = req.body.pimage;
        
         Category.findOne({title: title,_id: {$ne: id}},(err,category)=>{
             if (category){
-                res.redirect('/admin/category/edit-category/'+id)
+                fs.unlink('public/images/category-img/'+ pimage ,(err)=>{
+                    if(err) console.log(err);
+                    console.log('old img deleted');
+                });
+                req.flash('error', 'Category name already exists!');
+
+                return res.redirect('/admin/category/edit-category/'+id)
             }else{
-                Category.findByIdAndUpdate({_id:id},{$set:{title:title , image:req.file.filename}}).then((categories)=>{
+                Category.findByIdAndUpdate({_id:id},{$set:{title:title , image:image}}).then((category)=>{
                     category.save((err)=>{
                         if(err) return console.log(err);
+                        if (image !== ""){
+                            console.log('image is already there');
+
+                            if(category.image !== ""){
+                                console.log('image updated');
+
+                                fs.unlink('public/images/category-img/'+category.image ,(err)=>{
+                                    if(err) console.log(err);
+                                    console.log('old img deleted');
+
+                                })
+                            }
+
+                        }
+                        req.flash('success', `Category edited successfully!`);
+
                         res.redirect('/admin/category');
                     })
                 }).catch((err)=> console.log(err))
@@ -127,10 +168,37 @@ categoryRouter.post('/edit-category/:id',(req,res)=>{
 
 
 categoryRouter.get('/delete-category/:id',auth.isAdmin,(req,res)=>{
-    Category.findByIdAndRemove(req.params.id,(err)=>{
-        if(err) return console.log(err);
-        res.redirect('/admin/category');
-    });
+
+    Category.find({_id:req.params.id}, (err,category)=>{
+        console.log(category);
+        let cat = category[0].title;
+        console.log(cat);
+        Product.find({category:cat},(err,products)=>{
+            if(products.length > 0 ){
+                
+                console.log('products present');
+                console.log(products);
+                req.flash('error','there are some products in this category.Cant delete it.');
+                return res.redirect('/admin/category')
+            }else{
+                console.log('products absent');
+                Category.findByIdAndRemove(req.params.id,(err)=>{
+                        if(err) return console.log(err);
+                        req.flash('success', `Category deleted successfully!`);
+                
+                        res.redirect('/admin/category');
+                            });
+
+            }
+
+        })
+    })
+    // Category.findByIdAndRemove(req.params.id,(err)=>{
+    //     if(err) return console.log(err);
+    //     req.flash('success', `Category deleted successfully!`);
+
+    //     res.redirect('/admin/category');
+    // });
 });
